% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tables.R
\name{alignment_data}
\alias{alignment_data}
\title{alignment_data}
\usage{
alignment_data(data, table.width = 15, slice.position, return.data = FALSE)
}
\arguments{
\item{data}{A data.frame containing gene sequence data, which must contain the following columns:
\itemize{
\item{"position"}{ : sequence position, these must be sequential}
\item{"colour"}{ : colour for the sequence in the gene.alignment.tables visualisation}
\item{"..."}{ : any other columns will be displayed in the visualisation directly below the position number}
}}

\item{table.width}{the width of the gene.alignment.tables, defaults to 15 positions.}

\item{slice.position}{fii}
}
\description{
alignment_data converts a sequence dataset into the format required for the gene.alignment.tables visualisation
}
