% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tables.R
\name{alignment_DT}
\alias{alignment_DT}
\title{alignment_DT}
\usage{
alignment_DT(data, table.width = 15, slice.position, return.data = FALSE,
  alignment.table.id)
}
\arguments{
\item{data}{A data.frame with start-end pairs, needs the following columns
\itemize{
\item{"position"}{ : sequence position, these must be sequential}
\item{"colour"}{ : colour for the sequence in the gene.alignment.tables visualisation}
\item{"..."}{ : any other columns will be displayed in the visualisation directly below the position number}
}}

\item{table.width}{the width of the gene.alignment.tables, defaults to 20 positions.}

\item{slice.position}{fio}
}
\description{
alignment_DT generates a single datatable
}
