% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tables.R
\name{alignment_DT_unique_id}
\alias{alignment_DT_unique_id}
\title{alignment_DT_unique_id}
\usage{
alignment_DT_unique_id()
}
\description{
alignment_DT_unique_id generartes a unique ID for all datatable outputs shown within the sequence table visualisation
}
