% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coding_region_colours}
\alias{coding_region_colours}
\title{coding_region_colours}
\format{A data frame with 6 rows and 2 variables:
\describe{
\item{label}{Coding region name}
\item{colour}{Hex colour to use for the specified coding region}
}}
\source{
\url{http://philippa-matthews.com}
}
\usage{
coding_region_colours
}
\description{
Example colour codings for different coding regions found in both example datasets, \code{hbv_long_s_sequence} and \code{hbv_pol_sequence}.
}
\keyword{datasets}
