% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tables.R
\name{generate_dts}
\alias{generate_dts}
\title{generate_dts}
\usage{
generate_dts(data, table.width = 15, alignment.table.id)
}
\arguments{
\item{data}{A data.frame containing gene sequence data, which must contain the following columns:
\itemize{
\item{"position"}{ : sequence position, these must be sequential}
\item{"colour"}{ : colour for the sequence in the gene.alignment.tables visualisation}
\item{"..."}{ : any other columns will be displayed in the visualisation directly below the position number}
}}

\item{table.width}{the width of the gene.alignment.tables, defaults to 15 positions.}

\item{alignment.dt.unique.id}{The unique name given to DT::datatable objects created by generate_dts, must contain ONLY alphabetical characters. Use \code{alignment_DT_unique_id} to ensure a safe name is created.}
}
\description{
\code{generate_dts} generates a set of DT::datatable objects.
}
