% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hbv_pol_sequence}
\alias{hbv_pol_sequence}
\alias{hbv_pol_sequence}
\title{hbv_pol_sequence sequence data}
\format{A data frame with 1,143 rows and 6 variables:
\describe{
\item{position}{Gene position}
\item{reference}{Column to display}
\item{genom.a}{Alternative lettering}
\item{colour}{Colour to display this position as in the sequence viz}
}}
\source{
\url{http://philippa-matthews.com}

\url{http://philippa-matthews.com}
}
\usage{
hbv_pol_sequence

hbv_pol_sequence
}
\description{
An example sequence for HBV Pol

An example sequence for HBV Pol
}
\keyword{datasets}
