% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run gene.alignment.table examples}
\usage{
runExample()
}
\description{
Launch a \code{gene.alignment.tables} example Shiny app using the built-in
datasets.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  runExample()
}
}
