% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_legend.R
\name{coding_region_legend}
\alias{coding_region_legend}
\title{coding_region_legend}
\usage{
coding_region_legend(data)
}
\arguments{
\item{data}{A data.frame containing gene sequence data, which must contain the following columns:
\itemize{
\item{"label"}{ : label for the coding region}
\item{"colour"}{ : colour for the coding region}
}}
}
\description{
\code{coding_region_legend} creates a very simplistic legend of the coding regions using ggplot2.
}
